refactor(run): use dynamic import for subcommand
